vcl/gtk weld: Implement (un)select_all without forwarding
In the weld::ItemView subclasses in the vcl implementation
(SalInstanceItemView) and the gtk one (GtkInstanceTreeView
and GtkInstanceIconView), no longer let

* unselect_all() call select(-1)
* select_all() call unselect(-1)

and have logic in their weld::ItemView::select
and weld::ItemView::unselect overrides to select all or
clear selection all when called with those special indices.

Instead, let select() and unselect() call unselect_all()
and select_all() when called with the special indices,
i.e. forward the other way around and move the toolkit-specific
logic to (un)select all to the select_all()/unselect_all()
implementations.

This prepares for an upcoming commit that will
implement the index-based weld::ItemView::select
and weld::ItemView::unselect directly in the
weld::ItemView base class, i.e. the previous special
handling in the GTK and vcl implementations will
be dropped.

But the new logic to call (un)select_all() can be implemented
directly in the weld::ItemView base class as well.

No change in behavior intended.

Change-Id: I440cfa39fcc24231c6a72a67e079232c29c46900
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/195867
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>
Tested-by: Jenkins