tdf#168594 weld: Add IconView::get_item_at_rect
This method returns an iter to the item at the
given position.

Corresponding GTK API: gtk_icon_view_get_item_at_pos [1].
Qt also has similar API with QAbstractItemView::indexAt [2].

Initially, take over the logic already implemented in
LayoutMenu::CommandHdl (though the logic there uses indices
instead of iters) and use the newly added method in there.
(Using the mentioned GTK or Qt API for those VCL plugins
in the future could be an option.)

No change in behavior intended or seen in a quick test of
the scenario described in

    commit 3e0e0429aab01faafeb4faccbb5aa2800cbd9315
    Author: Michael Weghorn <m.weghorn@posteo.de>
    Date:   Wed Dec 17 17:15:08 2025 +0100

        weld: Add weld::IconView::get_rect taking iter param

Having the logic in weld::IconView allows it to be
reused from elsewhere in upcoming commits.

[1] https://docs.gtk.org/gtk3/method.IconView.get_item_at_pos.html
[2] https://doc.qt.io/qt-6/qabstractitemview.html#indexAt

Change-Id: I637c6c59263b0238ba21bfe39ea866666b964778
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/195870
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>
Tested-by: Jenkins