sd: Reuse weld::IconView::get_item_at_pos in MasterPagesSelector
Use the method newly introduced in previous commit

    Change-Id: I637c6c59263b0238ba21bfe39ea866666b964778
    Author: Michael Weghorn <m.weghorn@posteo.de>
    Date:   Thu Dec 18 21:26:41 2025 +0100

        tdf#168594  weld: Add IconView::get_item_at_rect

instead of implementing the same logic here again.

While the previous commit was for the IconView used
in Impress's "Properties" sidebar deck, this one is
for the "Master Slides" one and right-clicking on
one of the items opens a context menu as expected.

Support for opening the context menu using the keyboard
will be added separately.

Change-Id: I6a32e4a96f670312be19fe84cc3e9822e7ef7eb0
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/195871
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>