tdf#168594 weld: Add get_iter_index_in_parent for IconView
weld::TreeView already has such a method. Move it to
the weld::ItemView base class, to make it available
for weld::IconView as well.

The Qt and vcl TreeView implementations are fine for
IconView as well, so just move them to the corresponding
base classes.

For GTK, add a new
GtkInstanceIconView::get_iter_index_in_parent,
similar to the existing
GtkInstanceTreeView::get_iter_index_in_parent.

Change-Id: Ia22ac533b9d72b901eee9d17bdfe76c5738f871e
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/195891
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>