tdf#145759 Use DWMWA_USE_IMMERSIVE_DARK_MODE
Previously, magic number 20 was used for switching to dark mode on
Windows. On the other hand, DWMWA_USE_IMMERSIVE_DARK_MODE is defined for
that purpose since Windows 11 22000. Prior to that, it was an
undocumented constant:

DWMWINDOWATTRIBUTE enumeration (dwmapi.h)
https://learn.microsoft.com/en-us/windows/win32/api/dwmapi/ne-dwmapi-dwmwindowattribute

WinForms Dark title bar on Windows 10
https://stackoverflow.com/a/62811758

More informationa bout dark mode for win32 applications can be found in:

Support Dark and Light themes in Win32 apps
https://learn.microsoft.com/en-us/windows/apps/desktop/modernize/ui/apply-windows-themes

Change-Id: I96169ae87f4f42d6b0092128c527ba0e72f7529e
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/195910
Tested-by: Jenkins
Reviewed-by: Hossein <hossein@libreoffice.org>