tdf#168594 Change SvxCharView focus callback param to CharAndFont
What exact widget received focus is not of interest, but
only what special character is newly focused.

Therefore, switch the param of the handler from SvxCharView
to CharAndFont (and move that struct into the charwin.hxx
header, to be able to use it there).

This is one more step to prepare porting SfxCharmapCtrl away
from using SvxCharView in an upcoming commit.

Change-Id: I4b582e7b3e8564253f9016d8813adf6f6e25b191
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/195928
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>
Tested-by: Jenkins