tdf#168594 Only care about SvxCharView font family name
For the SvxCharView::maFont attributes, only the family
name is ever explicitly set from outside, see
SvxCharView::UpdateFont.

So from a semantic perspective, the widget holds a character's
text and font family.
Everything else are more or less internal implementation
details of how the widget displays a character itself.

Therefore, replace SvxCharWin::GetFont by a getter for
the font family, SvxCharWin::GetFontFamilyName.

Most of the callers of SvxCharview::GetFont also
only retrieve the font family from the font.

SvxCharacterMap::CharClickHdl is the only exception,
but there seems to be no reason for being interested
in anything else beyond the font family.
Adjust it to only explicitly set the same font family
for the other widgets in the "Special Characters"
dialog as well.

No obvious difference in behavior to before was noticed
with this commit in place when clicking on one of the
recent or favorite characters in the
"Insert" -> "Special Characters..." dialog.

This is one more step to prepare porting SfxCharmapCtrl away
from using SvxCharView in an upcoming commit.

Change-Id: I6189b7c1b78687994d09ee64e99d6c97eb4047ca
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/195931
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>
Tested-by: Jenkins