tdf#168594 Rework SvxCharView activation logic
SvxCharView is used in

1) the special characters toolbar popup
   (triggered by the "Insert Special Characters"
   toolbar button)

2) the special characters dialog
   (triggered via "Insert" -> "Special Character...")

In 1), a single click on a SvxCharView (i.e. a recent or
favorite character in the popup) triggers insertion into
the document and closes the popup.

In 2), a double-click triggers insertion into the document.

In both, 1) and 2), pressing the space or Enter key while
an SvxCharView has focus triggers the insertion into the
document.

This is implemented in different ways and places, partially
directly in SvxCharView, and partially in mouse click handlers.

Unify the approach by allowing to set a single activation handler,
and then making it the responsibility of that handler to insert
the character into the document and to do whatever else is needed
(closing the popup in case of 1).
This moves the char insertion logic out of SvxCharView.

SvxCharView::m_bActivateOnSingleClick now determines whether
the activation handler is called on a single mouse click
(true for the popup case, false for the dialog).

This also prepares for replacing the use of SvxCharView
in SfxCharmapContainer, by switching to a weld::IconView
instead.
SvxCharView::m_bActivateOnSingleClick can then be mapped to
the equivalent GtkIconView:activate-on-single-click [1]
property.

[1] https://docs.gtk.org/gtk3/property.IconView.activate-on-single-click.html

Change-Id: Ieadd263876b4e9ed0caf6aea45573655bacb4876
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/195940
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>