tdf#168594 a11y: Update special char info on keyboard triggered focus
So far, the information about the currently focused
recent/favorite character in the "Insert" -> "Special Character..."
dialog was only updated when a recent or favorite character
was clicked using the mouse, not when it received focus because
of moving focus to it using the Tab key.

Also update the info when focus is changed
using the keyboard, by connecting the handler to the
corresponding callback instead of the one specific
for mouse clicks.

Rename the handler for consistency.

Other than the special characters dialog, the
toolbar popup already implemented updating the
displayed info on focus change (see
SfxCharmapCtrl::CharFocusInHdl), so doesn't need
to be modified in that regard.

As a consequence of this, the last recent/favorite
character that had focus is now also used when
pressing the "Insert" button in the special characters
dialog, not necessarily the one that was previously
picked in the character table at the top.

This is consistent with the existing behavior when
a favorite/recent character was clicked on, so seems
fine to me.

Adjust the special character dialog a11y tests
accordingly, to press the Enter key while the
character to be inserted in the table is still
the active character. This triggers the dialog's
default action, which is the "Insert" one.

Change-Id: I0ff3067853a8f71f8ed7915f965ee6d78bfe8f8b
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/195933
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>
Tested-by: Jenkins