tdf#130857 tdf#168594 qt weld: Support activate-on-single-click
Both, GtkTreeView and GtkIconView have an
"activate-on-single-click" property
that specifies whether the "item-activated" signal
will be emitted after a single click. [1] [2]

For the Qt implementation, in case that property
is set to true, connect to the
QAbstractItemView::clicked signal that
"is emitted when a mouse button is left-clicked" [3]
and trigger the QAbstractItemView::activated [4]
signal that is handled in both, QtInstanceTreeView
and QtInstanceIconView to call the corresponding
weld::TreeView::signal_row_activated and
QtInstanceIconView::signal_item_activated.

In a quick test with a local change (s. below) to set
that property for Writer's "Insert" -> "Fontwork..."
dialog, single-clicking an item in that dialog now results
in it getting inserted into the document with the qt6
VCL plugin and SAL_VCL_QT_USE_WELDED_WIDGETS=1 .

Local test change to set the property for the dialog:

    diff --git a/svx/uiconfig/ui/fontworkgallerydialog.ui b/svx/uiconfig/ui/fontworkgallerydialog.ui
    index 65a749ab4a23..c7d649360694 100644
    --- a/svx/uiconfig/ui/fontworkgallerydialog.ui
    +++ b/svx/uiconfig/ui/fontworkgallerydialog.ui
    @@ -121,6 +121,7 @@
                         <property name="model">liststore1</property>
                         <property name="pixbuf-column">0</property>
                         <property name="tooltip-column">2</property>
    +                    <property name="activate-on-single-click">True</property>
                         <child internal-child="accessible">
                           <object class="AtkObject" id="ctlFavoriteswin-atkobject">
                             <property name="AtkObject::accessible-description" translatable="yes" context="fontworkgallerydialog|extended_tip|ctlFavoriteWin">Select a Fontwork style and click OK to insert the Fontwork into your document. Double-click or Ctrl (Command) + double-click the Fontwork in your document to enter text edit mode and change the text.</property>

[1] https://docs.gtk.org/gtk3/property.TreeView.activate-on-single-click.html
[2] https://docs.gtk.org/gtk3/property.IconView.activate-on-single-click.html
[3] https://doc.qt.io/qt-6/qabstractitemview.html#clicked
[4] https://doc.qt.io/qt-6/qabstractitemview.html#activated

Change-Id: I556cdbb4e17b3b9a864953cff39d1c477fb18098
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/196034
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>