tdf#168594 tdf#119931 a11y special chars: Use IconView for recent/favorites
For the recently used and favorite characters
in the special characters toolbar popup (triggered by
the "Insert Special Characters" toolbar button)
and the special characters dialog (triggered via
"Insert" -> "Special Character..."),
SfxCharmapContainer was so far using grids of the
custom SvxCharView, with each SvxCharView displaying
one character.

Switch from those grids of custom controls to
weld::IconView, one for the recent characters
and one for the favorites.

This ensures more standard/consistent interaction behavior,
like being able to navigate between the characters
in the IconView using the arrow keys (tdf#168594)
and allows dropping quite some custom logic.

Previous commits already reworked the logic/refactored
involved code to make the port to weld::IconView easier.

Some notes:

1) As described in more detail in previous commit

    commit 176f8acd7a993fc45ac7f027123270f80e4924ea
    Author: Michael Weghorn <m.weghorn@posteo.de>
    Date:   Fri Dec 19 19:52:45 2025 +0100

        tdf#168594 Rework SvxCharView activation logic

, the toolbar popup activates characters on single click
while the dialog does not. The
GtkIconView:activate-on-single-click [1] property is set
accordingly in the .ui file for the popup, but not for
the dialog.

2) The logic to draw the "icon" of a single character,
now located in SfxCharmapContainer::CreateIcon, is based
on the existing/previous logic from SvxCharView, previously
spread across multiple methods there (SvxCharView::SetDrawingArea,
SvxCharView::UpdateFont, SvxCharView::Paint, logic pretty much in
that order still).

3) When an item view loses focus, selection is cleared,
and it is restored when it receives focus again - see
ItemViewFocusInHdl and ItemViewFocusOutHdl. This
is because it is irrelevant what character was previously
selected in one of those icon views while interacting
with the other icon view or the bigger table of icons
in the special characters dialog.
(Still showing selection might be considered misleading.)

4) The switch to weld::IconView also addresses various of the
gla11y warnings in the special characters dialog
(related ticket: tdf#119931).

5) The previously used SvxCharView class is now unused, and
therefore dropped.

[1] https://docs.gtk.org/gtk3/property.IconView.activate-on-single-click.html

Change-Id: Id6407b1c4b154bc7004723a0534141107a6939d4
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/196035
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>