tdf#169610 dataprovider save,load SortTransformation
Before this patch, sorting could be applied in the dataprovider dialog,
but it was lost when saving the document. This patch adds saving and
loading of the SortTransformation.

The transformations of a dataprovider are applied to the data in a
clipboard document. Thus the range address in the SortParam starts
always in 0|0. That is different to the sorting that belongs to a
database range.

The file markup uses the elements table:sort and table:sort-by. But a
way different to a database-range is needed, because they are not
child elements of a <table:database-range>. The associated database
range is only indirectly known from the calcext:database-name
attribute of the calcext:data-mapping element.

Change-Id: I2425e4f7e342e2ab2895388c7ae08bd08afa6101
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/196037
Reviewed-by: Regina Henschel <rb.henschel@t-online.de>
Tested-by: Jenkins