weld: Deduplicate TreeView::get_cursor_index implementations
Instead of implementing it in every toolkit specific
subclass, this can be implemented in the abstract
base class by using the existing ItemView::get_cursor
method that returns a GtkTreeIter.

Also, move the method from weld::TreeView to the weld::ItemView
base class, as it is applicable for the weld::IconView subclass
in the same way.

Change-Id: I907f6890de433634fba1b3fc5788add520c9ae46
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/196056
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>