tdf#168751 sw interdependent redlines, format on delete: fix unwanted combine
The DOCX bugdoc has a format redline (mark as bold) + a delete redline,
and these two overlap. Open the document in Writer, reject the delete
redline, the boldness remains in the document, which is unexpected.

The problem is an in case a format redline is created and that would be
on top of an insert or delete from the same author, then we currently
merge this into the insert/delete redline. The reason for this seems to
be that search&replace inserts text to the document and then copies
direct format from the insert position to the inserted text range, and
this format should not show up on the UI, since the UI action was just
to replace text. Another variant of this is autocorrect "make first
character of the paragraph uppercase" in action, which is internally
again a replace.

Fix the problem by disabling combine of redlines in
sw::DocumentRedlineManager::PreAppendFormatRedline(), in case we would
do a combine and this is a format on delete. This way the format is
rejected by the UI reject action and the unwanted format on insert still
doesn't show up. The behavior was like this already in case the authors
of the format and the delete redlines were different.

Note that this tweak is not even important for the insert-then-format
case: accept doesn't modify the body text and reject would remove text
from the body text, so we get rid of the unwanted format anyway.

Change-Id: Ib08b0c62c012f14801deae47a4b85d77e32fc1a8
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/196066
Tested-by: Jenkins CollaboraOffice <jenkinscollaboraoffice@gmail.com>
Reviewed-by: Mike Kaganski <mike.kaganski@collabora.com>
(cherry picked from commit de12d41f140aee260fa751cd75b19ce3dc895dbf)
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/196533
Reviewed-by: Caolán McNamara <caolan.mcnamara@collabora.com>
Tested-by: Caolán McNamara <caolan.mcnamara@collabora.com>