weld: Deduplicate TreeView::get_text implementations
The variant taking an int param can be implemented using
the variant taking a GtkTreeIter param.

Do so once in the abstract weld::TreeView base class, and drop
the toolkit-specific implementations.

Also move the method declarations of both methods
next to each other, so one comment can describe both
of them (and there's no need to keep that in sync
if anything changes in the future).

Change-Id: Ifc6f2211e6ec1fda36a7e6f355edf09d3dc1d020
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/196129
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>
Tested-by: Jenkins