tdf#170003 writerfilter: no belowSpacing emulation with tables
This patch fixes my 25.2.6
commit 1326e09d019f05a82265f15c26288b4ffb7dc0c2
tdf#167657 writerfilter: only move sectPr bottomMargin after pageBreak

It turns out that m_xPreStartingRange is a bit of a disaster.
First of all, it seems to be set even if gotoPreviousParagraph fails.
That can't be a good thing.

Secondly (and the problem for me), it jumps over tables,
so if the page break happens after a table
then I was applying the below spacing to an earlier paragraph.

make CppunitTest_sw_ooxmlexport25 \
    CPPUNIT_TEST_NAME=testTdf170003_bottomSpacing

Change-Id: I883dd397413001fafb000e1e90c0e6738b3b4c87
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/196143
Reviewed-by: Justin Luth <jluth@mail.com>
Tested-by: Jenkins
(cherry picked from commit 82936569e0e17705b6de501d7151549c51520fb9)
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/196246