tdf#170095 override soft edge if there are 3D effects
In OOXML, 3D effects override the soft-edge effect. Currently in LO 3D
effects are not implemented, they are just remembered in a grabbag.

To bring LO's behavior closer to other software, if when loaded there
are both 3D effects and soft edge effect we ignore the soft edge effect.
And if, while editing, a user adds soft edge effect then we dumb the
remembered 3D effects in the grabbag.

Change-Id: I2f59422d25c24bb4f9042694e0315bb15a2ac4c5
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/196151
Tested-by: Jenkins CollaboraOffice <jenkinscollaboraoffice@gmail.com>
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>