tdf#152706: don't overwrite source on translation failure
linguistic::Translate returns an empty string, when any error occurs
(it may be a malformed URL, or wrong key, or a server error). Then
PasteHTMLToPaM would replace the selection with nothing, and after
that, show a "Requested clipboard format is not available" message.

This change checks the returned string, to avoid the replacement. It
doesn't show a message itself (TODO in a follow-up).

Change-Id: Ia92b0d0a609386bf3988fc9ef7e01da84311ab38
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/196253
Reviewed-by: Mike Kaganski <mike.kaganski@collabora.com>
Tested-by: Jenkins
Signed-off-by: Xisco Fauli <xiscofauli@libreoffice.org>
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/196280