Add -nologo to linker invocation in MSVC
Avoids a console noise from linker in VS2026, which shows the
logo and then the whole output, somehow not filtered out by
filter-creatingLibrary.awk for that version without the switch.

Amazingly, the switch cannot appear before -LIB (it will give

  LINK : warning LNK4044: unrecognized option '/LIB'; ignored

and then LNK1120 error). And it cannot appear too late in the
command line: it would be ignored there. Sigh.

Change-Id: Ifa4b0c0d55f2ba703ed8b77339729466117ee3bf
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/196187
Tested-by: Jenkins
Reviewed-by: Mike Kaganski <mike.kaganski@collabora.com>