sw: Simplifying the "Insert Table" dialog
The table creation dialog currently contains some
confusing code related to converting a style index
in the tree to a style index in the table. A magic
number of 255 has been introduced, which has
no connection to the objective limitations of the
style list and requires separate processing when
defining a style.

This change merges the "- None -" style and the
document's table styles into a single list. The
"- None -" style is placed before all others,
preserving the current behavior. This also
removes the artificial limitation of 255 styles,
expanding it to the maximum of size_t.

Change-Id: Iac3be58d4473b184742fe5cb32ca4f1240f3d805
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/196203
Tested-by: Jenkins
Reviewed-by: Mike Kaganski <mike.kaganski@collabora.com>