tdf#170149 Fix clicking on a character in the input method's popup window
[self setMarkedText:selectedRange:replacementRange:] does not
get called on macOS Tahoe. So when pressing and holding a key,
the original character would never get selected and clicking
on a replacement character in the input method's popup window
would fail. Surprisingly, pressing a replacement character's
number still worked.

Also, make the following scenarios behave as close as possible
to the TextEdit application whether or not there is any arrowing
into the input method's popup window. I tested with Canadian and
Japanese - Romaji keyboards:
- Pressing the Escape or Return keys commits uncommitted text
- Pressing the Delete key discards uncommitted text
- Pressing the Backspace or Fn-Delete keys commits uncommitted text

Change-Id: I7c3fdfecb4d1979d38e9f652bea8c7d66e6c9612
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/196239
Tested-by: Jenkins
Reviewed-by: Patrick Luby <guibomacdev@gmail.com>
Signed-off-by: Xisco Fauli <xiscofauli@libreoffice.org>
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/196858
Reviewed-by: Christian Lohmaier <lohmaier+LibreOffice@googlemail.com>