vcl: use frozen in OpenType feature definition list
We can now simplify this list and use frozen containers as there is
no more parameters for OpenType features. If we will add parameters,
we can still add those as exceptions, no need to add them in the
list itself.

Singleton OpenTypeFeatureDefinitionListPrivate now doesn't make
much sense as the methods became static, so the whole class was
removed and only functions are used.

Change-Id: Icc66cdf0a5eb4bd0997b4815421c044372dfd395
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/196273
Tested-by: Jenkins
Reviewed-by: Tomaž Vajngerl <quikee@gmail.com>